Components are places into appropriate folders
