--- src/components/Gamebox.css
@@ -1,3 +1,0 @@
-html, body {
-    overflow: hidden;
-  }
