+++ src/components/game/Gamebox.js
