+++ src/pages/Rules.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Carousel from '../components/rules/Carousel';
+
+export default class Rules extends React.Component {
+  render() {
+    return (
+      <div>
+        <Carousel />
+      </div>
+    )
+  }
+}
